# swagger-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "New.Zcdn WebApi 接口文档    .NET 10.0.7",
    "description": "Build: 2026-09-20 15:53:25",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/tokenRefreshes": {
      "post": {
        "tags": ["授权管理"],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/New.Zcdn.SharedModel.Dto.Core.Policy.PolicyRefreshTokenDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/New.Zcdn.SharedModel.Dto.Core.Policy.PolicyRefreshTokenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/New.Zcdn.SharedModel.Dto.Core.Policy.PolicyRefreshTokenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/New.Zcdn.SharedModel.Dto.Core.Policy.PolicyRefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "New.Zcdn.SharedModel.Dto.Core.Policy.PolicyRefreshTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "刷新令牌。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "刷新政策大脑访问令牌。"
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "apiKey",
        "description": "JWT授权(数据将在请求头中进行传输) 直接在下框中输入Bearer {token} \"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}